refactor: API 구조 변경 및 GitHub 서비스 변경
